+++ instrumentation/opentelemetry-instrumentation-dbapi/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+- bugfix: cursors and connections now produce spans when used with context managers
+  ([#1028](https://github.com/open-telemetry/opentelemetry-python/pull/1028))
+
+## Version 0.12b0
+
+Released 2020-08-14
+
 - Change package name to opentelemetry-instrumentation-dbapi
   ([#966](https://github.com/open-telemetry/opentelemetry-python/pull/966))