+++ instrumentation/opentelemetry-instrumentation-requests/CHANGELOG.md
@@ -2,8 +2,17 @@
 
 ## Unreleased
 
+- Add support for instrumenting prepared requests
+  ([#1040](https://github.com/open-telemetry/opentelemetry-python/pull/1040))
+
+## Version 0.12b0
+
+Released 2020-08-14
+
 - Change package name to opentelemetry-instrumentation-requests
   ([#961](https://github.com/open-telemetry/opentelemetry-python/pull/961))
+- Span name reported updated to follow semantic conventions to reduce
+  cardinality ([#972](https://github.com/open-telemetry/opentelemetry-python/pull/972))
 
 ## 0.7b1